chore(local): catch bazel-do issues before push
