+++ dev/ci/internal/ci/BUILD.bazel
@@ -26,6 +26,7 @@ go_library(
     visibility = ["//dev/ci:__subpackages__"],
     deps = [
         "//dev/ci/gitops",
+        "//dev/ci/helpers",
         "//dev/ci/images",
         "//dev/ci/internal/buildkite",
         "//dev/ci/internal/ci/changed",
@@ -34,7 +35,6 @@ go_library(
         "//dev/sg/root",
         "//internal/lazyregexp",
         "//lib/errors",
-        "@com_github_grafana_regexp//:regexp",
         "@com_github_masterminds_semver//:semver",
         "@com_github_sourcegraph_log//:log",
         "@in_gopkg_yaml_v2//:yaml_v2",
@@ -45,7 +45,6 @@ go_test(
     name = "ci_test",
     srcs = [
         "aspect_workflows_test.go",
-        "bazel_operations_test.go",
         "wolfi_operations_test.go",
     ],
     data = [
